requirements.txt
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
openpyxl>=3.1.0
pandas>=2.0.0
requests>=2.31.0
geopy>=2.4.0
h3>=3.7.6
Pillow>=10.0.0